horas en uso subido

+++ app/src/main/java/com/iticbcn/trokka/UsageInfo.kt
@@ -1,7 +1,7 @@
 package com.iticbcn.trokka
 
 data class UsageInfo (
-    var horasUso: Long = 0,
+    var horasUso: Double = 0.0,
     var countVisitasMapa: Int = 0,
     var countEntradesApp: Int = 0,  // Veces que entra a la app
     var countCrearObjeto: Int = 0,
